Support deferred plugin configuration
In mysql 8.0, plugins are disabled during mysqld --initialize. This has
the effect of making plugin configuration in the my.cnf invalid during
mysqld --initialize.  This was observed in pxc-release by initializing
with audit log support enabled leading to this error:

  > [ERROR] [MY-000067] [Server] unknown variable 'audit-log-file=foo'.

To support this, plugins should now be configured under a
[mysqld_plugin] section which is only processed during a normal mysqld
startup.