Fix and simplify smoke-tests packaging
Smoke tests package one of the integration test suites.  This is failing
because src/specs/vendor/ was removed in an earlier commit but the
packaging script explicitly set the go build -mod=vendor flag.

Made some changes to simplify the build script to avoid a direct ginkgo
cli dependency and not dependend on vendoring.